test: separate tests into different files
