+++ contracts/hub/meta_transaction/src/lib.rs
@@ -6,9 +6,6 @@ pub mod migrate;
 pub mod query;
 pub mod state;
 
-#[cfg(test)]
-mod tests;
-
 #[cfg(not(target_arch = "wasm32"))]
 mod interface;
 #[cfg(not(target_arch = "wasm32"))]
